fix(tres): allow five minute strict wake arming

+++ board/drivers/drivers.h
@@ -63,7 +63,7 @@ extern volatile bool wake_monitor_som_off_seen;
 extern volatile bool wake_monitor_som_off_ready;
 extern volatile uint8_t wake_monitor_som_off_countdown;
 extern volatile bool wake_monitor_can_armed;
-extern volatile uint8_t wake_monitor_can_baseline_countdown;
+extern volatile uint16_t wake_monitor_can_baseline_countdown;
 extern volatile int8_t wake_monitor_tesla_counter;
 extern volatile int8_t wake_monitor_tesla_door_counter;
 extern volatile uint8_t wake_monitor_tesla_front_door_known_mask;
